Event handling moved out from main module
